style: rename folder name to *s
