--- src/store/reducers/index.ts
@@ -1,10 +1,0 @@
-import { combineReducers } from '@reduxjs/toolkit'
-import mainSlice from './main'
-
-const rootReducers = combineReducers({
-  main: mainSlice,
-})
-
-export default rootReducers
-
-export { updateCounter, increaseCounter, decreaseCounter } from './main'
