+++ src/views/home/index.tsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from 'react-redux'
 import { PlusOutlined } from '@ant-design/icons'
 import { Icon } from '@iconify/react'
 import { Main } from '@/components'
-import { updateCounter, type AppDispatch, type RootState } from '@/store'
+import { updateCounter, type AppDispatch, type RootState } from '@/stores'
 import styles from './index.module.css'
 
 const Home: React.FC = () => {
